initial deploy to heroku

+++ src/database/database.provider.ts
@@ -27,14 +27,16 @@ export const databaseProviders = [
         seeds: [ResourceSeeder, CitySeeder, DistrictSeeder, SubDistrictSeeder],
         factories: [],
       };
-      await dropDatabase({
-        options: options as DataSourceOptions,
-        ifExist: true,
-      });
-      await createDatabase({
-        options: options as DataSourceOptions,
-        ifNotExist: true,
-      });
+      if (process.env.APP_ENV === 'development') {
+        await dropDatabase({
+          options: options as DataSourceOptions,
+          ifExist: true,
+        });
+        await createDatabase({
+          options: options as DataSourceOptions,
+          ifNotExist: true,
+        });
+      }
       const dataSource = new DataSource(options);
       const source = await dataSource.initialize();
       await runSeeders(dataSource);
